+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/eventstory/eventstory/adapter/EventFeedListAdapter.kt
@@ -1,42 +1,23 @@
 package com.teameetmeet.meetmeet.presentation.eventstory.eventstory.adapter
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import com.teameetmeet.meetmeet.data.model.Feed
-import com.teameetmeet.meetmeet.databinding.ItemEventFeedBinding
 import com.teameetmeet.meetmeet.presentation.eventstory.eventstory.OnFeedItemClickListener
 
 class EventFeedListAdapter(
     private val onItemClickListener: OnFeedItemClickListener
-) : ListAdapter<Feed, EventFeedListAdapter.EventFeedViewHolder>(diffCallback) {
+) : ListAdapter<Feed, EventFeedViewHolder>(diffCallback) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): EventFeedViewHolder {
-        val binding = ItemEventFeedBinding.inflate(
-            LayoutInflater.from(parent.context),
-            parent,
-            false
-        )
-        return EventFeedViewHolder(binding, onItemClickListener)
+        return EventFeedViewHolder.from(parent, onItemClickListener)
     }
 
     override fun onBindViewHolder(holder: EventFeedViewHolder, position: Int) {
         holder.bind(getItem(position))
     }
 
-    class EventFeedViewHolder(
-        private val binding: ItemEventFeedBinding,
-        private val onItemClickListener: OnFeedItemClickListener
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: Feed) {
-            binding.item = item
-            itemView.setOnClickListener {
-                onItemClickListener.onItemClick(item)
-            }
-        }
-    }
 
     companion object {
         private val diffCallback = object : DiffUtil.ItemCallback<Feed>() {
